requirements.txt
chardet
pandas
pyyaml
docker-compose